+++ app/src/main/java/com/github/bumblebee202111/doubean/data/repository/TvRepository.kt
@@ -1,9 +1,9 @@
 package com.github.bumblebee202111.doubean.data.repository
 
 import com.github.bumblebee202111.doubean.network.ApiService
-import com.github.bumblebee202111.doubean.network.model.NetworkTvDetail
+import com.github.bumblebee202111.doubean.network.model.subject.NetworkTvDetail
 import com.github.bumblebee202111.doubean.network.model.toPhotoList
-import com.github.bumblebee202111.doubean.network.model.toTvDetail
+import com.github.bumblebee202111.doubean.network.model.subject.toTvDetail
 import com.github.bumblebee202111.doubean.network.util.makeApiCall
 import javax.inject.Inject
 import javax.inject.Singleton